feat: four turbulence rows, and two bugs where this UI quietly disagreed with the solver
The solver grew turbulence models. This is the UI catching up, which took
about forty lines. The rest of this commit is two bugs I fell over on the way
past and could not walk away from, both of the same shape: this UI holding an
opinion about the solver's output that the solver did not share.

=== THE EASY HALF ===

    TURBULENCE  Turbulence model        none | smagorinsky | kOmegaSST
                Smagorinsky Cs          the one constant smagorinsky has
                Turbulence intensity    fraction of the inlet speed
                Turbulence length       m, the biggest eddy coming in

Four rows, three of them read by one model each. `Cs` only goes out for
smagorinsky. The two inlet rows only go out for kOmegaSST. `none` sends the
model name and nothing else, so turning the group off costs exactly one key
on the command line and changes nothing about what the solver does.

Held back on finding `turbLengthScale` in the executable, like every block
since 0.2. Point it at a 0.7 solver and none of the four rows reach the
command line and it gets asked exactly what it was asked before.

Refused before launching rather than after: a Cs of zero or above one, an
intensity above one (an inlet more turbulent than it is moving), a negative
length scale, and a model name this solver has never heard of.

=== NOW THE HALF THAT IS EMBARRASSING ===

**No two-fluid frame has ever opened in this UI.** Not one. Since 0.5.

I was making a phase frame to check something unrelated, dropped it on the
reader, and got

    Unknown scalar array before required arrays: phase

The reader refuses any array it does not recognise until it has seen all three
of pressure, mask and velocity. The solver writes the phase fraction BETWEEN
the pressure and the mask. So every frame the entire multiphase feature has
ever produced was rejected on the way in, and the branch that shipped it
never once loaded its own output.

k and omega go in the same place, so turbulent frames would have gone the same
way, which is how I found it - two features' worth of frames failing on one
line of somebody's caution.

The order check is deleted. Where a scalar sits in a file is the writer's
business. A frame that never provides the three required arrays is still
rejected, at the end of the parse, which is the only place that can actually
be known - and that check was already there, doing the job, the whole time.

VtkFrameTests now writes a frame with a scalar before the mask and insists it
comes back with the right values in it, so this cannot quietly return.

=== WHAT YOU GET OUT OF IT ===

`phase` shows up in the Field button. So do `nuT`, `wallDistance` and `strain`
when the solver was asked for them, and `k` and `omega`, which arrive on their
own in every kOmegaSST frame because the solver needs them back to continue.
None of that needed new code - the registry has always walked whatever turned
up. It just needed the frames to get through the door.

=== AND THE OTHER ONE, WHICH IS SUBTLER AND FUNNIER ===

"Adapter verification failed: Fluid Solver VTK mask differs from the GUI
preview." Three cells out of 2500. On a cube.

The UI cuts the section itself, writes `section-adapter.obj`, and then checks
the mask the solver rasterised out of that file against the one it drew for
the preview. Both rasterisers use the same rule - a cell is on the boundary if
its centre is within one cell circumradius of the outline - so they should
agree exactly.

They do not, because **the solver keeps its grid spacing in float and the
preview computed it in double.** `float(1)/50` is 0.019999999552965164.
`1.0/50` is 0.02. By i = 30 the two disagree about where the cell centre is by
about one part in 10^8.

One part in 10^8 is nothing. It is nothing right up until a face of the model
lands exactly on a cell boundary, which for anything axis aligned it does ON
PURPOSE. Then the four corner cells of the rectangle sit at exactly one
circumradius from the outline - precisely the distance the rule turns on - and
one part in 10^8 is the entire difference between a solid cell and an empty
one.

Fix: the preview divides in float as well. Being more accurate than the thing
you are previewing is not an improvement, it is a different answer.

The result is not symmetrical and that took a second look too. float(1)/50 is
a hair UNDER 0.02, so a cell centre creeps towards the origin as its index
grows: the far corners land just inside the circumradius, the near ones just
outside. Three corners solid, one not. Looks like a bug. Is the solver's own
answer. A preview is not allowed to have a nicer one.

Checked on twelve shape/grid/angle combinations, including two disconnected
contours and rotated slices: every cell identical, and the real UI now says
"Fluid Solver mask matches the GUI preview" where it used to refuse.

**Nothing in the solver changed for this.** The solver was right; the preview
was showing you a run you were not going to get.

=== ALSO, WHILE I WAS IN THERE ===

FluidSolverRunTests grew the BODIES coverage that the 0.7 commit message
claimed it had. It did not have it. The message said "FluidSolverRunTests
covers the block end to end" and the file contained the string "body" exactly
zero times. That is on me. It covers it now: the block held back from a solver
that has never heard of it, everything sent to one that has, a bounciness sent
only when collisions are on, a coupling that does not exist refused, a
bounciness above one refused, a multi-line bodyMotion refused, and a body told
to travel through an empty domain refused.

Plus the same for turbulence, and the four-key contract for the base solver
is still 24 arguments.

=== CHECKED, BY CLICKING ON IT ===

Virtual display, synthetic mouse events, same as always. Set the model row to
kOmegaSST and the intensity to 0.14, hit Run, and the run's own
requested-arguments.txt says

    turbulence=kOmegaSST
    turbIntensity=0.14388489723205566
    turbLengthScale=0

with no `Cs=` anywhere, which is exactly right for that model. Then opened the
frames it wrote, walked the Field button round to nuT, and looked at the eddy
viscosity sitting where it should be - against the walls and around the block.
Same run also came back with "Fluid Solver mask matches the GUI preview",
which is a sentence this UI had never said about that model.

All four suites pass. GeometryProcessorTests pins all four corner cells of a
square whose faces land on cell boundaries, so a well meaning double division
cannot creep back in.

+++ src/GeometryProcessor.cpp
@@ -1015,8 +1015,11 @@ MaskResult GeometryProcessor::rasterizeContours(
     }
     result.contours = std::move(contours);
     result.cells.assign(width * height, 0);
-    const double dx = parameters.Lx / parameters.nx;
-    const double dy = parameters.Ly / parameters.ny;
+
+    const double dx = static_cast<float>(parameters.Lx) /
+                      static_cast<float>(parameters.nx);
+    const double dy = static_cast<float>(parameters.Ly) /
+                      static_cast<float>(parameters.ny);
     const double boundaryRadius = 0.5 * std::hypot(dx, dy);
 
     for (int j = 0; j < parameters.ny; ++j) {
